refactor: 엔티티 구조로 리팩토링 및 불필요한 파일 삭제
