+++ app/blog/app/posts/[id]/page.tsx
@@ -1,11 +1,12 @@
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 
-import { getPostBySlug, markdownToHtml } from '@lib';
+import { getPostBySlug } from '@entities/posts/api';
+import { getReadingTime } from '@entities/posts/lib/calculateReadingTime';
+import { markdownToHtml } from '@shared/lib/markdownToHtml';
+import { BLOG_URL } from '@shared/constants/url';
+import { NICKNAME } from '@shared/constants/nickname';
 import { Content, PostHead } from './components';
-import { BLOG_URL } from '@constants/url';
-import { NICKNAME } from '@constants/nickname';
-import { getReadingTime } from '@utils';
 
 interface PostProps {
   params: Promise<{
