+++ app/blog/src/entities/posts/api/index.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import matter from 'gray-matter';
 import { join } from 'path';
 
-import { Post } from '@interface/post';
+import { Post } from '../model/types';
 
 const postsDirectory = join(process.cwd(), 'posts');
