+++ app/blog/src/entities/posts/lib/calculateReadingTime.ts
