+++ app/blog/src/entities/posts/model/types.ts
