--- app/blog/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './api';
-export { markdownToHtml } from './markdownToHtml';
