+++ app/blog/src/shared/types/global.d.ts
