--- app/blog/src/utils/index.ts
@@ -1,1 +1,0 @@
-export { getReadingTime } from './calculateReadingTime';
